feat(composition): integrate Autofac for desktop service management and validation

+++ tests/ChapterTool.Core.Tests/Importing/IndexImporterTests.cs
@@ -352,7 +352,7 @@ public sealed class IndexImporterTests
         builder.Reserved(UserDataPadding);
 
         builder.SeekTo(indexesAddress);
-        builder.UInt32BE((uint)IndexesMinContentSize);
+        builder.UInt32BE(IndexesMinContentSize);
         builder.Reserved(TitleEntrySize);
         builder.Reserved(TitleEntrySize);
         builder.UInt16BE(0);
